Add TransCanada module version (incompatible with current code base)

+++ DSCResources/cTentacleAgent/cTentacleAgent.psm1
@@ -12,14 +12,15 @@ function Get-TargetResource
         [ValidateSet("Started", "Stopped")]
         [string]$State = "Started",
         
+        [bool]$AutoRegister,
         [string]$ApiKey,
         [string]$OctopusServerUrl,
         [string[]]$Environments,
         [string[]]$Roles,
         [string]$DefaultApplicationDirectory,
-        [int]$ListenPort,
-        [string]$RegisteredNic,
-        [bool]$isNatted=$false
+        [int]$ListenPort,        
+        [string]$OctopusHomeDirectory,
+        [string]$OctopusServerThumbprint
     )
 
     Write-Verbose "Checking if Tentacle is installed"
@@ -73,24 +74,20 @@ function Set-TargetResource
         [ValidateSet("Started", "Stopped")]
         [string]$State = "Started",
         
+        [bool]$AutoRegister = $true,
         [string]$ApiKey,
         [string]$OctopusServerUrl,
         [string[]]$Environments,
         [string[]]$Roles,
         [string]$DefaultApplicationDirectory = "$($env:SystemDrive)\Applications",
         [int]$ListenPort = 10933,
-        [string]$RegisteredNic,
-        [bool]$isNatted=$false
-   )
+        [string]$OctopusHomeDirectory = "$($env:SystemDrive)\Octopus",
+        [string]$OctopusServerThumbprint        
+    )
 
     if ($Ensure -eq "Absent" -and $State -eq "Started") 
     {
         throw "Invalid configuration: service cannot be both 'Absent' and 'Started'"
-    }
-
-    if ( (-not $InitialDeploy) -and ($DeployProject -or $DeployVersion))
-    {
-        throw "Invalid configuration: Resource set to not do initial deploy but Project and/or Version to deploy to specified"
     }
 
     $currentResource = (Get-TargetResource -Name $Name)
@@ -106,7 +103,9 @@ function Set-TargetResource
 
     if ($Ensure -eq "Absent" -and $currentResource["Ensure"] -eq "Present")
     {
-        Remove-TentacleRegistration -name $Name -apiKey $ApiKey -octopusServerUrl $OctopusServerUrl
+        if ($AutoRegister) {
+            Remove-TentacleRegistration -name $Name -apiKey $ApiKey -octopusServerUrl $OctopusServerUrl
+        }
         
         $serviceName = (Get-TentacleServiceName $Name)
         Write-Verbose "Deleting service $serviceName..."
@@ -114,8 +113,8 @@ function Set-TargetResource
         
         # Uninstall msi
         Write-Verbose "Uninstalling Tentacle..."
-        $tentaclePath = "$($env:SystemDrive)\Octopus\Tentacle.msi"
-        $msiLog = "$($env:SystemDrive)\Octopus\Tentacle.msi.uninstall.log"
+        $tentaclePath = "$OctopusHomeDirectory\Tentacle.msi"
+        $msiLog = "$OctopusHomeDirectory\Tentacle.msi.uninstall.log"
         if (test-path $tentaclePath)
         {
             $msiExitCode = (Start-Process -FilePath "msiexec.exe" -ArgumentList "/x $tentaclePath /quiet /l*v $msiLog" -Wait -Passthru).ExitCode
@@ -133,7 +132,9 @@ function Set-TargetResource
     elseif ($Ensure -eq "Present" -and $currentResource["Ensure"] -eq "Absent") 
     {
         Write-Verbose "Installing Tentacle..."
-        New-Tentacle -name $Name -apiKey $ApiKey -octopusServerUrl $OctopusServerUrl -port $ListenPort -RegisteredNic $RegisteredNic -isNatted $isNatted -environments $Environments -roles $Roles -DefaultApplicationDirectory $DefaultApplicationDirectory
+        New-Tentacle -name $Name -AutoRegister $AutoRegister -apiKey $ApiKey -octopusServerUrl $OctopusServerUrl `
+            -port $ListenPort -environments $Environments -roles $Roles -DefaultApplicationDirectory $DefaultApplicationDirectory `
+            -OctopusHomeDirectory $OctopusHomeDirectory -OctopusServerThumbprint $OctopusServerThumbprint
         Write-Verbose "Tentacle installed!"
     }
 
@@ -160,14 +161,15 @@ function Test-TargetResource
         [ValidateSet("Started", "Stopped")]
         [string]$State = "Started",
         
+        [bool]$AutoRegister,
         [string]$ApiKey,
         [string]$OctopusServerUrl,
         [string[]]$Environments,
         [string[]]$Roles,
         [string]$DefaultApplicationDirectory,
-        [int]$ListenPort,
-        [string]$RegisteredNic,
-        [bool]$isNatted=$false
+        [int]$ListenPort,    
+        [string]$OctopusHomeDirectory,
+        [string]$OctopusServerThumbprint
     )
  
     $currentResource = (Get-TargetResource -Name $Name)
@@ -228,34 +230,12 @@ function Invoke-AndAssert {
 # After the Tentacle is registered with Octopus, Tentacle listens on a TCP port, and Octopus connects to it. The Octopus server
 # needs to know the public IP address to use to connect to this Tentacle instance. Is there a way in Windows Azure in which we can 
 # know the public IP/host name of the current machine?
-function Get-MyPublicIPAddress([string]$RegisteredNic,[bool]$isNatted,[string]$OctopusServerUrl,[int]$port){
-    #First Verify the adapter exists
-    $netAdapter = Get-NetAdapter -InterfaceAlias $RegisteredNic -ErrorAction SilentlyContinue
-    if($netAdapter -eq $null){
-        throw "Selected NIC $($RegisteredNic) does not exist"
-    }
-
-    #If adapter is natted, find the actual Public IP
-    if($isNatted){
-        Write-Verbose "NIC $($RegisteredNic) is natted. Determining actual public IP"
-        $downloader = new-object System.Net.WebClient
-        $ip = $downloader.DownloadString("http://icanhazip.com").Trim()
-    }
-    else{
-        Write-Verbose "Getting IP address of $($RegisteredNic) NIC"
-        $ip = Get-NetIPAddress -InterfaceAlias $RegisteredNic -AddressFamily IPv4 | select -exp IPAddress
-    }
-
-    #Test the connection to the Octopus Server. If it is not reachable, throw an error
-    $urlRegex = ‘([a-zA-Z]{3,})://([\w-]+\.)+[\w-]+(/[\w- ./?%&=]*)*?’
-    if($OctopusServerUrl -match $urlRegex){
-        $OctopusServerUrl = $OctopusServerUrl.Split("//") | select -Last 1
-    }
-    $adapterTest = Test-NetConnection $OctopusServerUrl -Port $port
-    if(!($adapterTest.TcpTestSucceeded -and ($adapterTest.InterfaceAlias -eq $RegisteredNic))){
-        throw "Cannot reach Octopus Server $($OctopusServerUrl) from Network $($RegisteredNic). Please check your connection and try running your configuration again"
-    }
-    else{return $ip}
+function Get-MyPublicIPAddress
+{
+    Write-Verbose "Getting public IP address"
+    $downloader = new-object System.Net.WebClient
+    $ip = $downloader.DownloadString("http://ifconfig.me/ip")
+    return $ip
 }
  
 function New-Tentacle 
@@ -264,35 +244,48 @@ function New-Tentacle
         [Parameter(Mandatory=$True)]
         [string]$name,
         [Parameter(Mandatory=$True)]
+        [bool]$AutoRegister,
         [string]$apiKey,
-        [Parameter(Mandatory=$True)]
         [string]$octopusServerUrl,
-        [Parameter(Mandatory=$True)]
         [string[]]$environments,
-        [Parameter(Mandatory=$True)]
         [string[]]$roles,
         [int] $port,
-        [string]$RegisteredNic,
-        [bool]$isNatted=$false,
-        [string]$DefaultApplicationDirectory
+        [Parameter(Mandatory=$True)]
+        [string]$DefaultApplicationDirectory,
+        [Parameter(Mandatory=$True)]
+        [string]$OctopusHomeDirectory,
+        [string]$OctopusServerThumbprint
     )
  
     if ($port -eq 0) 
     {
         $port = 10933
     }
-
+    if ($AutoRegister -and (!$apiKey -or !$octopusServerUrl -or !$environments -or !$roles)) 
+    {
+        throw "When creating Tentacle using AutoRegister=true the follow parameters must be provided: apiKey, octopusServerUrl, environments, roles"
+    }
+    if (!$AutoRegister -and !$OctopusServerThumbprint) 
+    {
+        throw "When creating Tentacle using AutoRegister=false the follow parameters must be provided: OctopusServerThumbprint"
+    }
+    
     Write-Verbose "Beginning Tentacle installation" 
   
-    $tentacleDownloadUrl = "http://octopusdeploy.com/downloads/latest/OctopusTentacle64"
+    #$tentacleDownloadUrl = "http://octopusdeploy.com/downloads/latest/OctopusTentacle64"
+    # Temporarily using version 2.6 instead to be compatible with 2.x server
+    $tentacleDownloadUrl = "https://download.octopusdeploy.com/octopus/Octopus.Tentacle.2.6.5.1010-x64.msi"
+            
     if ([IntPtr]::Size -eq 4) 
     {
-        $tentacleDownloadUrl = "http://octopusdeploy.com/downloads/latest/OctopusTentacle"
+        #$tentacleDownloadUrl = "http://octopusdeploy.com/downloads/latest/OctopusTentacle"
+        # Temporarily using version 2.6 instead to be compatible with 2.x server
+        $tentacleDownloadUrl = "https://download.octopusdeploy.com/octopus/Octopus.Tentacle.2.6.5.1010.msi"
     }
 
-    mkdir "$($env:SystemDrive)\Octopus" -ErrorAction SilentlyContinue
+    mkdir $OctopusHomeDirectory -ErrorAction SilentlyContinue
 
-    $tentaclePath = "$($env:SystemDrive)\Octopus\Tentacle.msi"
+    $tentaclePath = "$OctopusHomeDirectory\Tentacle.msi"
     if ((test-path $tentaclePath) -ne $true) 
     {
         Write-Verbose "Downloading latest Octopus Tentacle MSI from $tentacleDownloadUrl to $tentaclePath"
@@ -300,7 +293,7 @@ function New-Tentacle
     }
   
     Write-Verbose "Installing MSI..."
-    $msiLog = "$($env:SystemDrive)\Octopus\Tentacle.msi.log"
+    $msiLog = "$OctopusHomeDirectory\Tentacle.msi.log"
     $msiExitCode = (Start-Process -FilePath "msiexec.exe" -ArgumentList "/i $tentaclePath /quiet /l*v $msiLog" -Wait -Passthru).ExitCode
     Write-Verbose "Tentacle MSI installer returned exit code $msiExitCode"
     if ($msiExitCode -ne 0) 
@@ -308,48 +301,62 @@ function New-Tentacle
         throw "Installation of the Tentacle MSI failed; MSIEXEC exited with code: $msiExitCode. View the log at $msiLog"
     }
  
-    Write-Verbose "Open port $port on Windows Firewall"
-    Invoke-AndAssert { & netsh.exe advfirewall firewall add rule protocol=TCP dir=in localport=$port action=allow name="Octopus Tentacle: $Name" }
-    
-    $ipAddress = Get-MyPublicIPAddress -RegisteredNic $RegisteredNic -isNatted $isNatted -OctopusServerUrl $octopusServerUrl -port $port
- 
-    Write-Verbose "Public IP address: $ipAddress"
-    Write-Verbose "Configuring and registering Tentacle"
+    $windowsFirewall = Get-Service -Name MpsSvc   
+    if ($windowsFirewall.Status -eq "Running")
+    {
+        Write-Verbose "Open port $port on Windows Firewall"
+        Invoke-AndAssert { & netsh.exe advfirewall firewall add rule protocol=TCP dir=in localport=$port action=allow name="Octopus Tentacle: $Name" }
+    }
+    else
+    {
+        Write-Verbose "Windows Firewall Service is not running... skipping firewall rule addition"
+    }
+            
+    Write-Verbose "Configuring Tentacle"
   
     pushd "${env:ProgramFiles}\Octopus Deploy\Tentacle"
- 
-    $tentacleHomeDirectory = "$($env:SystemDrive)\Octopus"
-    $tentacleAppDirectory = $DefaultApplicationDirectory
-    $tentacleConfigFile = "$($env:SystemDrive)\Octopus\$Name\Tentacle.config"
-    Invoke-AndAssert { & .\tentacle.exe create-instance --instance $name --config $tentacleConfigFile --console }
-    Invoke-AndAssert { & .\tentacle.exe configure --instance $name --home $tentacleHomeDirectory --console }
-    Invoke-AndAssert { & .\tentacle.exe configure --instance $name --app $tentacleAppDirectory --console }
+
+    Invoke-AndAssert { & .\tentacle.exe create-instance --instance $name --config "$OctopusHomeDirectory\$Name\Tentacle.config" --console }
+    Invoke-AndAssert { & .\tentacle.exe configure --instance $name --home $OctopusHomeDirectory --console }
+    Invoke-AndAssert { & .\tentacle.exe configure --instance $name --app $DefaultApplicationDirectory --console }
     Invoke-AndAssert { & .\tentacle.exe configure --instance $name --port $port --console }
+    if (!$AutoRegister) 
+    {
+        Write-Verbose "Adding trusted server thumbprint $OctopusServerThumbprint"
+        Invoke-AndAssert { & .\tentacle.exe configure --instance $Name --trust $OctopusServerThumbprint --console }
+    }
     Invoke-AndAssert { & .\tentacle.exe new-certificate --instance $name --console }
     Invoke-AndAssert { & .\tentacle.exe service --install --instance $name --console }
 
-    $registerArguments = @("register-with", "--instance", $name, "--server", $octopusServerUrl, "--name", $env:COMPUTERNAME, "--publicHostName", $ipAddress, "--apiKey", $apiKey, "--comms-style", "TentaclePassive", "--force", "--console")
-
-    foreach ($environment in $environments) 
+    if ($AutoRegister) 
     {
-        foreach ($e2 in $environment.Split(',')) 
+        $ipAddress = Get-MyPublicIPAddress
+        $ipAddress = $ipAddress.Trim()
+        Write-Verbose "Public IP address: $ipAddress"
+        Write-Verbose "Automatically registering Tentacle with server $octopusServerUrl"
+        $registerArguments = @("register-with", "--instance", $name, "--server", $octopusServerUrl, "--name", $env:COMPUTERNAME, "--publicHostName", $ipAddress, "--apiKey", $apiKey, "--comms-style", "TentaclePassive", "--force", "--console")
+    
+        foreach ($environment in $environments) 
         {
-            $registerArguments += "--environment"
-            $registerArguments += $e2.Trim()
+            foreach ($e2 in $environment.Split(',')) 
+            {
+                $registerArguments += "--environment"
+                $registerArguments += $e2.Trim()
+            }
         }
-    }
-    foreach ($role in $roles) 
-    {
-        foreach ($r2 in $role.Split(',')) 
+        foreach ($role in $roles) 
         {
-            $registerArguments += "--role"
-            $registerArguments += $r2.Trim()
+            foreach ($r2 in $role.Split(',')) 
+            {
+                $registerArguments += "--role"
+                $registerArguments += $r2.Trim()
+            }
         }
+    
+        Write-Verbose "Registering with arguments: $registerArguments"
+        Invoke-AndAssert { & .\tentacle.exe ($registerArguments) }
     }
-
-    Write-Verbose "Registering with arguments: $registerArguments"
-    Invoke-AndAssert { & .\tentacle.exe ($registerArguments) }
-
+    
     popd
     Write-Verbose "Tentacle commands complete"
 }
@@ -380,3 +387,4 @@ function Remove-TentacleRegistration
         Write-Verbose "Could not find Tentacle.exe"
     }
 }
+
